move previous TimeRangeSelector logic to TimeRangeControls in dashboards package
